requirements.txt
streamlit
pandas
plotly
pytest
openpyxl
fpdf2
kaleido
